pyproject.toml
[project]
name = "order-platform-msa-infer-pipeline"
version = "0.1.0"
description = "Inference service for order platform forecasting (FastAPI + AutoGluon)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.0.6",
    "autogluon-timeseries",
    "dvc>=3.63.0",
    "mlflow>=3.3.2",
    "boto3",
    "s3fs",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "notebook>=7.4.5",
    "pandas-stubs==2.3.2.250827",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "ruff>=0.12.11",
    "nbformat>=5.10.4",
    "plotly>=6.3.0",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "ipykernel>=6.30.1",
]


